+++ src/builders/utils.rs
@@ -17,14 +17,16 @@ pub fn generate_rss_channel(opts: &Options, feeds: &Vec<Feed>) -> String {
         .description(format!("Aggregated Liveboat rss feed for {}", opts.title))
         .build();
     let mut items = Vec::new();
+    let mut article_guids: Vec<i64> = Vec::new();
     for feed in feeds {
-        if feed.is_query_feed() || feed.is_empty() {
-            continue;
-        }
         let mut fc = feed.clone();
         fc.truncate_items();
 
         for feed_item in fc.items {
+            if article_guids.iter().any(|i| i == &feed_item.guid()) {
+                continue;
+            };
+            article_guids.push(feed_item.guid().clone());
             items.push(feed_item)
         }
     }
@@ -173,7 +175,58 @@ mod tests {
         f3.update_with_url_data(Vec::new(), true, None, 1);
 
         let result = generate_rss_channel(&Options::default(), &Vec::from([f1, f2, f3]));
-        assert_eq!(result,  "<?xml version=\"1.0\" encoding=\"utf-8\"?><rss version=\"2.0\"><channel><title>Liveboat feed page</title><link></link><description>Aggregated Liveboat rss feed for Liveboat feed page</description><item><title>item3</title><link>http://test3.com</link><author>exaroth</author><pubDate>Tue, 3 Dec 2024 04:26:40 +0000</pubDate></item><item><title>item2</title><link>http://test2.com</link><author>exaroth</author><pubDate>Mon, 2 Dec 2024 00:40:00 +0000</pubDate></item><item><title>item1</title><link>http://test1.com</link><author>exaroth</author><pubDate>Sat, 30 Nov 2024 20:53:20 +0000</pubDate></item><item><title>item5</title><link>http://test5.com</link><author>exaroth</author><pubDate>Sat, 30 Nov 2024 20:53:20 +0000</pubDate></item></channel></rss>")
+        assert_eq!(result,  "<?xml version=\"1.0\" encoding=\"utf-8\"?><rss version=\"2.0\"><channel><title>Liveboat feed page</title><link></link><description>Aggregated Liveboat rss feed for Liveboat feed page</description><item><title>item3</title><link>http://test3.com</link><author>exaroth</author><pubDate>Tue, 3 Dec 2024 04:26:40 +0000</pubDate></item><item><title>item2</title><link>http://test2.com</link><author>exaroth</author><pubDate>Mon, 2 Dec 2024 00:40:00 +0000</pubDate></item><item><title>item1</title><link>http://test1.com</link><author>exaroth</author><pubDate>Sat, 30 Nov 2024 20:53:20 +0000</pubDate></item><item><title>item4</title><link>http://test4.com</link><author>exaroth</author><pubDate>Sat, 30 Nov 2024 20:53:20 +0000</pubDate></item><item><title>item5</title><link>http://test5.com</link><author>exaroth</author><pubDate>Sat, 30 Nov 2024 20:53:20 +0000</pubDate></item></channel></rss>")
+    }
+
+    #[test]
+    fn test_generating_rss_with_query_feeds() {
+        let mut f = Feed::init(
+            "www.example.com/rss".to_string(),
+            "Test feed".to_string(),
+            "www.example.com".to_string(),
+        );
+        let i1 = FeedItem::new(
+            "item1",
+            "http://test1.com",
+            "",
+            "exaroth",
+            1733000000,
+            false,
+            "Test content 1",
+            1,
+        );
+        f.items.push(i1.clone());
+        let mut qf = Feed::init_query_feed(String::from("Test query feed"), 1);
+        qf.items.push(i1.clone());
+        let result = generate_rss_channel(&Options::default(), &Vec::from([f, qf]));
+        assert_eq!(result,  "<?xml version=\"1.0\" encoding=\"utf-8\"?><rss version=\"2.0\"><channel><title>Liveboat feed page</title><link></link><description>Aggregated Liveboat rss feed for Liveboat feed page</description><item><title>item1</title><link>http://test1.com</link><author>exaroth</author><pubDate>Sat, 30 Nov 2024 20:53:20 +0000</pubDate></item></channel></rss>")
+    }
+    #[test]
+    fn test_generting_rss_with_hidden_feeds() {
+        let mut f = Feed::init(
+            "www.example.com/rss".to_string(),
+            "Test feed".to_string(),
+            "www.example.com".to_string(),
+        );
+        f.update_with_url_data(
+            Vec::new(),
+            true,
+            None,
+            1,
+        );
+        let i1 = FeedItem::new(
+            "item1",
+            "http://test1.com",
+            "",
+            "exaroth",
+            1733000000,
+            false,
+            "Test content 1",
+            1,
+        );
+        f.items.push(i1);
+        let result = generate_rss_channel(&Options::default(), &Vec::from([f]));
+        assert_eq!(result,  "<?xml version=\"1.0\" encoding=\"utf-8\"?><rss version=\"2.0\"><channel><title>Liveboat feed page</title><link></link><description>Aggregated Liveboat rss feed for Liveboat feed page</description><item><title>item1</title><link>http://test1.com</link><author>exaroth</author><pubDate>Sat, 30 Nov 2024 20:53:20 +0000</pubDate></item></channel></rss>")
     }
 
     #[test]